+++ lib/cli/option.dart
@@ -18,7 +18,7 @@ abstract class Flag extends _BaseOption<bool> {
           null,
           null,
           defaultsTo,
-          callback == null ? null : (value) => callback(value as bool),
+          callback == null ? null : (value) => callback(value),
           OptionType.flag,
           negatable: negatable,
           hide: hide,
@@ -56,21 +56,6 @@ abstract class Option extends _BaseOption<String> {
 }
 
 abstract class _BaseOption<T> {
-  final String name;
-  final String? abbr;
-  final String? help;
-  final String? valueHelp;
-  final Iterable<String>? allowed;
-  final Map<String, String>? allowedHelp;
-  final T? defaultsTo;
-  final Function(T?)? callback;
-  final OptionType type;
-  final bool negatable;
-  final bool? splitCommas;
-  final bool mandatory;
-  final bool hide;
-  final List<String> aliases;
-
   _BaseOption(
     this.name,
     this.abbr,
@@ -87,4 +72,19 @@ abstract class _BaseOption<T> {
     this.hide = false,
     this.aliases = const [],
   });
+
+  final String name;
+  final String? abbr;
+  final String? help;
+  final String? valueHelp;
+  final Iterable<String>? allowed;
+  final Map<String, String>? allowedHelp;
+  final T? defaultsTo;
+  final void Function(T?)? callback;
+  final OptionType type;
+  final bool negatable;
+  final bool? splitCommas;
+  final bool mandatory;
+  final bool hide;
+  final List<String> aliases;
 }