+++ lib/data/keys.dart
@@ -1,0 +1,1 @@
+const elixirKey = 'GENERATED_BY_ELIXIR';