+++ lib/usecase/comment_indicates.dart
@@ -1,35 +1,72 @@
+import 'package:elixir/data/keys.dart';
+import 'package:elixir/data/model/comment.dart';
 import 'package:elixir/data/model/indicate.dart';
-import 'package:elixir/data/model/review_comment.dart';
 import 'package:elixir/data/repository/github.dart';
+import 'package:elixir/extension/iterable.dart';
+import 'package:meta/meta.dart';
 
 class CommentIndicatesUseCase {
-  final GitHubRepository _gitHubRepository;
-
   const CommentIndicatesUseCase({
-    GitHubRepository gitHubRepository = const GitHubRepository(),
+    required GitHubRepository gitHubRepository,
   }) : _gitHubRepository = gitHubRepository;
 
-  Future<List<ReviewComment>> call({
+  final GitHubRepository _gitHubRepository;
+
+  Future<String> call({
     required String repo,
     required int num,
     required Iterable<Indicate> indicates,
   }) async {
-    final comments = <ReviewComment>[];
-
-    // TODO: use https://docs.github.com/ja/rest/pulls/reviews#create-a-review-for-a-pull-request
-    for (final indicate in indicates) {
-      final comment = await _gitHubRepository.createReviewComment(
-        repo: repo,
-        num: num,
-        body: indicate.body,
-        commitId: indicate.commitId,
+    final commentKeyGroup = indicates.groupBy<_CommentKey>(
+      (indicate) => _CommentKey(
         path: indicate.path,
         line: indicate.line,
-      );
-      comments.add(comment);
-      await Future.delayed(Duration(seconds: 10));
-    }
+      ),
+    );
 
-    return comments;
+    final comments = commentKeyGroup.entries.map((entry) {
+      final key = entry.key;
+      final value = entry.value;
+      final body = '''
+<!-- $elixirKey -->
+<table>
+  <tbody>
+${value.map((e) => e.message).join('\n')}
+  </tbody>
+</table>''';
+      return Comment(
+        path: key.path,
+        line: key.line,
+        body: body,
+      );
+    }).toList();
+
+    return _gitHubRepository.createReview(
+      repo: repo,
+      num: num,
+      comments: comments,
+    );
   }
+}
+
+@immutable
+class _CommentKey {
+  const _CommentKey({
+    required this.path,
+    required this.line,
+  });
+
+  final String path;
+  final int line;
+
+  @override
+  bool operator ==(Object other) =>
+      identical(this, other) ||
+      other is _CommentKey &&
+          runtimeType == other.runtimeType &&
+          path == other.path &&
+          line == other.line;
+
+  @override
+  int get hashCode => path.hashCode ^ line.hashCode;
 }