+++ test/data/repository/dart_test.dart
@@ -1,10 +1,11 @@
 import 'dart:io';
 
-import 'package:mockito/annotations.dart';
-import 'package:mockito/mockito.dart';
-import 'package:elixir/data/model/analysis.dart';
+import 'package:elixir/data/model/analyze_result.dart';
 import 'package:elixir/data/repository/dart.dart';
 import 'package:elixir/data/source/process.dart';
+import 'package:mockito/annotations.dart';
+import 'package:mockito/mockito.dart';
+import 'package:path/path.dart';
 import 'package:test/test.dart';
 
 import 'dart_test.mocks.dart';
@@ -15,110 +16,76 @@ Future<void> main() async {
   final repository = DartRepository(runner: runner);
   final dir = Directory('./');
 
-  test('Diagnostics is empty.', () async {
-    when(runner.run(any, ignoreError: anyNamed('ignoreError'))).thenAnswer((_) async => '''
-Analyzing flutter-mobile-lints...      0.8s
-No issues found!''');
-
-    final expected = AnalysisResults(
-      version: 0,
-      diagnostics: <Diagnostic>[],
-    );
-
+  test('Diagnostic is empty.', () async {
+    when(runner.run(any, ignoreError: anyNamed('ignoreError')))
+        .thenAnswer((_) async => '');
+    const expected = Iterable<AnalyzeResult>.empty();
     final actual = await repository.analyze(dir: dir);
     expect(actual, expected);
   });
 
   test('One diagnostic exists.', () async {
-    when(runner.run(any, ignoreError: anyNamed('ignoreError'))).thenAnswer((_) async => '''
-Analyzing danger-flutter-lint-demo...  1.7s
-{"version":1,"diagnostics":[{"code":"argument_type_not_assignable","severity":"ERROR","type":"COMPILE_TIME_ERROR","location":{"file":"/Users/blendthink/Documents/repositories/danger-flutter-lint-demo/lib/main.dart","range":{"start":{"offset":305,"line":14,"column":14},"end":{"offset":309,"line":14,"column":18}}},"problemMessage":"The argument type 'Null' can't be assigned to the parameter type 'String'.","documentation":"https://dart.dev/diagnostics/argument_type_not_assignable"}]}''');
-
-    final expected = AnalysisResults(
-      version: 1,
-      diagnostics: <Diagnostic>[
-        Diagnostic(
-          code: 'argument_type_not_assignable',
-          severity: DiagnosticSeverity.error,
-          type: DiagnosticType.compileTimeError,
-          location: Location(
-            file:
-                '/Users/blendthink/Documents/repositories/danger-flutter-lint-demo/lib/main.dart',
-            range: Range(
-              start: Position(offset: 305, line: 14, column: 14),
-              end: Position(offset: 309, line: 14, column: 18),
-            ),
-          ),
-          problemMessage:
-              "The argument type 'Null' can't be assigned to the parameter type 'String'.",
-          documentation:
-              'https://dart.dev/diagnostics/argument_type_not_assignable',
-        ),
-      ],
+    when(runner.run(any, ignoreError: anyNamed('ignoreError'))).thenAnswer(
+      (_) async =>
+          '''ERROR|SYNTACTIC_ERROR|EXPECTED_TOKEN|$current/lib/main.dart|5|23|1|Expected to find ';'.''',
     );
-
+    const expected = <AnalyzeResult>[
+      AnalyzeResult(
+        severity: Severity.error,
+        type: 'SYNTACTIC_ERROR',
+        errorCode: 'EXPECTED_TOKEN',
+        filePath: 'lib/main.dart',
+        line: 5,
+        column: 23,
+        length: 1,
+        errorMessage: """Expected to find ';'.""",
+      ),
+    ];
     final actual = await repository.analyze(dir: dir);
     expect(actual, expected);
   });
 
   test('Multiple diagnostics exist.', () async {
-    when(runner.run(any, ignoreError: anyNamed('ignoreError'))).thenAnswer((_) async => '''
-Analyzing danger-flutter-lint-demo...  1.7s
-{"version":1,"diagnostics":[{"code":"argument_type_not_assignable","severity":"ERROR","type":"COMPILE_TIME_ERROR","location":{"file":"/Users/blendthink/Documents/repositories/danger-flutter-lint-demo/lib/main.dart","range":{"start":{"offset":305,"line":14,"column":14},"end":{"offset":309,"line":14,"column":18}}},"problemMessage":"The argument type 'Null' can't be assigned to the parameter type 'String'.","documentation":"https://dart.dev/diagnostics/argument_type_not_assignable"},{"code":"unrecognized_error_code","severity":"WARNING","type":"STATIC_WARNING","location":{"file":"/Users/blendthink/Documents/repositories/danger-flutter-lint-demo/analysis_options.yaml","range":{"start":{"offset":121,"line":5,"column":5},"end":{"offset":125,"line":5,"column":9}}},"problemMessage":"'demo' isn't a recognized error code."},{"code":"always_put_required_named_parameters_first","severity":"INFO","type":"LINT","location":{"file":"/Users/blendthink/Documents/repositories/danger-flutter-lint-demo/lib/main.dart","range":{"start":{"offset":1094,"line":33,"column":45},"end":{"offset":1099,"line":33,"column":50}}},"problemMessage":"Put required named parameters first.","documentation":"https://dart-lang.github.io/linter/lints/always_put_required_named_parameters_first.html"}]}''');
-
-    final expected = AnalysisResults(
-      version: 1,
-      diagnostics: <Diagnostic>[
-        Diagnostic(
-          code: 'argument_type_not_assignable',
-          severity: DiagnosticSeverity.error,
-          type: DiagnosticType.compileTimeError,
-          location: Location(
-            file:
-                '/Users/blendthink/Documents/repositories/danger-flutter-lint-demo/lib/main.dart',
-            range: Range(
-              start: Position(offset: 305, line: 14, column: 14),
-              end: Position(offset: 309, line: 14, column: 18),
-            ),
-          ),
-          problemMessage:
-              "The argument type 'Null' can't be assigned to the parameter type 'String'.",
-          documentation:
-              'https://dart.dev/diagnostics/argument_type_not_assignable',
-        ),
-        Diagnostic(
-          code: 'unrecognized_error_code',
-          severity: DiagnosticSeverity.warning,
-          type: DiagnosticType.staticWarning,
-          location: Location(
-            file:
-                '/Users/blendthink/Documents/repositories/danger-flutter-lint-demo/analysis_options.yaml',
-            range: Range(
-              start: Position(offset: 121, line: 5, column: 5),
-              end: Position(offset: 125, line: 5, column: 9),
-            ),
-          ),
-          problemMessage: "'demo' isn't a recognized error code.",
-        ),
-        Diagnostic(
-          code: 'always_put_required_named_parameters_first',
-          severity: DiagnosticSeverity.info,
-          type: DiagnosticType.lint,
-          location: Location(
-            file:
-                '/Users/blendthink/Documents/repositories/danger-flutter-lint-demo/lib/main.dart',
-            range: Range(
-              start: Position(offset: 1094, line: 33, column: 45),
-              end: Position(offset: 1099, line: 33, column: 50),
-            ),
-          ),
-          problemMessage: 'Put required named parameters first.',
-          documentation:
-              'https://dart-lang.github.io/linter/lints/always_put_required_named_parameters_first.html',
-        ),
-      ],
+    when(runner.run(any, ignoreError: anyNamed('ignoreError'))).thenAnswer(
+      (_) async => '''
+ERROR|SYNTACTIC_ERROR|EXPECTED_TOKEN|$current/lib/main.dart|5|23|1|Expected to find ';'.
+WARNING|STATIC_WARNING|UNSUPPORTED_OPTION_WITH_LEGAL_VALUE|$current/analysis_options.yaml|4|3|3|The option 'foo' isn't supported by 'linter'. Try using the only supported option: ''rules''.
+INFO|HINT|UNUSED_LOCAL_VARIABLE|$current/lib/main.dart|4|9|3|The value of the local variable 'foo' isn't used.''',
     );
 
+    const expected = <AnalyzeResult>[
+      AnalyzeResult(
+        severity: Severity.error,
+        type: 'SYNTACTIC_ERROR',
+        errorCode: 'EXPECTED_TOKEN',
+        filePath: 'lib/main.dart',
+        line: 5,
+        column: 23,
+        length: 1,
+        errorMessage: """Expected to find ';'.""",
+      ),
+      AnalyzeResult(
+        severity: Severity.warning,
+        type: 'STATIC_WARNING',
+        errorCode: 'UNSUPPORTED_OPTION_WITH_LEGAL_VALUE',
+        filePath: 'analysis_options.yaml',
+        line: 4,
+        column: 3,
+        length: 3,
+        errorMessage:
+            """The option 'foo' isn't supported by 'linter'. Try using the only supported option: ''rules''.""",
+      ),
+      AnalyzeResult(
+        severity: Severity.info,
+        type: 'HINT',
+        errorCode: 'UNUSED_LOCAL_VARIABLE',
+        filePath: 'lib/main.dart',
+        line: 4,
+        column: 9,
+        length: 3,
+        errorMessage: """The value of the local variable 'foo' isn't used.""",
+      ),
+    ];
     final actual = await repository.analyze(dir: dir);
     expect(actual, expected);
   });